Ranked Choice Voting and Political Affiliation
 In this project, I will go over how political affiliation specifically Democrats,Republicans and Independents feel about Ranked Choice Voting. According to Time Magazine, Ranked-choice voting is an electoral system that involves allowing people to vote for more than one candidate. Instead of normal voting, people will be allowed to have a first, second, and third choice for each position they are voting for.

The focus of the project will be analyzing data to conclude if there is a significant and noticeable different in how Democrats, Republicans, and Independents view ranked-choice voting. If RCV is implemented this would help in swing states and close competitive elections. We will be using a machine learning algorithm for this project which is a Decision Tree Regression. First we will be using a hypothesis test to statistically test which political affiliation is more prone to RCV. Then we will be using a regression decision tree to test our data to see a true significance.